+++ src/Dictionary.php
@@ -6,7 +6,6 @@ namespace Bakame\Http\StructuredFields;
 
 use Iterator;
 use Throwable;
-use TypeError;
 use function array_key_exists;
 use function array_keys;
 use function array_map;
@@ -259,7 +258,7 @@ final class Dictionary implements OrderedMap
     {
         return match (true) {
             $member instanceof InnerList, $member instanceof Item => self::filterForbiddenState($member),
-            $member instanceof StructuredField => throw new TypeError('Expecting a "'.Item::class.'" or a "'.InnerList::class.'" instance; received a "'.$member::class.'" instead.'),
+            $member instanceof StructuredField => throw new InvalidArgument('Expecting a "'.Item::class.'" or a "'.InnerList::class.'" instance; received a "'.$member::class.'" instead.'),
             default => Item::from($member),
         };
     }