+++ src/InnerList.php
@@ -6,7 +6,6 @@ namespace Bakame\Http\StructuredFields;
 
 use Iterator;
 use Throwable;
-use TypeError;
 use function array_filter;
 use function array_map;
 use function array_splice;
@@ -51,7 +50,7 @@ final class InnerList implements MemberList, ParameterAccess
     {
         return match (true) {
             $member instanceof Item => self::filterForbiddenState($member),
-            $member instanceof StructuredField => throw new TypeError('Expecting a "'.Item::class.'" instance; received a "'.$member::class.'" instance instead.'),
+            $member instanceof StructuredField => throw new InvalidArgument('Expecting a "'.Item::class.'" instance; received a "'.$member::class.'" instance instead.'),
             default => Item::from($member),
         };
     }